requirements.txt
requests==2.28.1
python-dotenv==0.21.0
pyyaml==6.0.0